Updated to meet Acceptance Criteria and allow user to choose output format
In format 1 the program works as initially.

In format 2: the output file has an extensionof .csv, the output file name is the title of the board without spaces, the content of the file is structured from left to right and is delimited by commas, and only content with at least 1 vote is included.